requirements.txt
django-cms>=2.3
django-easy-select2>=1.2.0